Use latest git-resource to bypass certificate issue
Getting packages from gopkg.in was throwing certificate issues.
This was caused by a combination of:
- Letsencrypt changing its root certificate authority
- git resource-type being pinned to a specific image by default

Old image for git resource doesn't include the new authority.
The problem can easily be solved by redefining the git resource
type to use latest official image.

In a future Concourse vers. this change shouldn't be necessary.