Use the LTS version of OpenLDAP
As of 2022-10-24, the LTS branch of OpenLDAP is 2.5. The
latest features branch is 2.6.  Also, the latest 2.5 will
build on xenial, and the latest 2.6 will not. We should jump
to a later branch when we stop supporting xenial, or when 2.5
stops being supported.

See https://www.openldap.org/software/download/ for which
branches are currently supported.

[#183555103]