make pipeline more reliable
the docker based bosh started in a privileged container is causing the pipeline
to flake. The stemcell download has been failing repeatedly when the deployment
against this director

switch the task to run against a smith env. This env will be claimed later on in
the pipeline anyway. This should also speed up the whole pipeline as we do not
need to wait for the docker bosh to be created.

fly-by:
- update to ginkgo v2