test against real tas bosh
the docker bosh for running the bosh release tests used an outdated
very old xenial stemcell (stuck at 621.125) and created flakes in CI.

This removes the docker cpi in favor of a real bosh env provided by shepherd.

fly-by: add explicit input of mapfs. the bosh release tests required them to be
able to run, previously the old task used the copy contained in one of the submodules.