add deploy test for upgrade in place
mapfs-release is broken under certain bosh deploy conditions.

the release starting with v1.2.53 have a bug in the pre-start
script that prevents the pre-start script to work when:
- the underlying vm didn't change
- the mapfs release did change
then following conditions would be met:
- /lib/libfuse3.so exists

from running again
when the release is upgraded without an accompanying recreation
of the VM.

If a bosh deploy command will not recreate the VM when applying the
new version of mapfs-release from above affected versions, then the
pre start script will follow a code path where it tries to recreate
a symlink that already exists and fails.

This should onlt