remove harbor
[#187074138]

it's going away and we're in the process of moving away from it.

- ensure to always use cryogenics-essentials provided by resource

additionally we decided to stop bumping task images with cryogenics/essentials
versions. Instead we want to use concourse a resource in combination with the `image`
attribute of the pipeine task step to manage the version of cryo essentials that a
task is using.

to enable this approach we need to override each task with the appropriate image
provided by the get step at the beginning of each job.