remove usage of nested harbor hosted image
[#187378303]

the task from mapfs-release-concourse-tasks/scripts/ci/run_bosh_release_tests.build.yml
will try to run another docker container ( it wants to execute the ginkgo suite ) against
a bosh running in a dind setup. The 2nd image is required to have out test / build
tools available and is currently pulled from harbor.

to avoid authentication issues and rate limits, this updates the pipeline to provide
an image.tar that can be imported into the dind setup of this task so it needs no
direct pull access to dockerhub or any other registry.